core/requirements.txt
asgiref==3.7.2
Django==4.2.15
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
gunicorn==21.2.0
